requirements.txt
anthropic>=0.25.0
streamlit>=1.35.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.7.0
plotly>=5.20.0
python-dotenv>=1.0.0
groq>=0.9.0
ollama>=0.2.0
pymupdf>=1.24.0
python-docx>=1.1.0
openpyxl>=3.1.0
